new cross and display plus multiple choises
adding the new cross methos.
adding a display with turtle or matplotlib for the bee path.
adding the possibility to choose between the two cross methodes, the choise to use the check for convergence, the choise to display the best bee path using turtle and the choise to display using matplotlib and puting the display fonction in a new file named graphic.py